requirements.txt

requests==2.32.5
python-dotenv==1.1.1
geocoder==1.38.1
